apply review : divide classes for StringCalculator
